Added enum Status to CitybikesStation
It is useful for asking to the station the status of the free bikes or
the available places based on thresholds set in the constructor.
It has been though when the station has to be added to the map with a
color different for the status of the station, i.e. red for NO_FREE,
yellow for LOW_FREE and green for MEDIUM_FREE.