Fixed some bugs and style corrections
Added a new AlertDialog in case of no station have been found for the city the user is in,
added log call for better debug diagnostics, moved some code to
functions for better reading. Added new strings for the new AlertDialog.
Fixed ConcurrentModificationException when removing an item to a Map
during iteration.