Reorganization of the code
Added new info for the infobox.
Added Resources reference to the MainActivity.
Moved centre of the map to onChangeLocation() of
LocationBroadcastReceiver. Now it is working as expected.
Added Intent filter for location updates.
Refactor of some code and simplyfied onResume() since it did same things
of onStart().
Added a new progress state and new strings for the infobox to show.