Added code for use the overlay inside the map.
Make use of the reference lastOverlay that make easier for the user to
make changes at a single overlay without passing its name every time in
every function.
Added several functions for controlling the visibility of an overlay or
a list of overlays.
Made a single ItemizedIconOverlay.OnItemGestureListener which is called
rather frequently inside the code.