Takes in account that the user can be in a town where there are no stations
I added a new method to OSMNominatimService class that gets the
display_name field from the JSON from the OSM Nominatim web service.
I moved the block that takes care of the station maps in MainActivity to
the top branch of the if.
I moved the common re-enable things on the UI (like buttons to refresh
map) to the common part of the onPostExecute, where common means that
re-enabling operations are part of the code that take cares of stations
as well as code that doesn't.